+++ include/guppy.h
@@ -1,6 +1,6 @@
 /*
  *	guppy.h -- Guppy language interface for nocc
- *	Copyright (C) 2010-2015 Fred Barnes <frmb@kent.ac.uk>
+ *	Copyright (C) 2010-2016 Fred Barnes <frmb@kent.ac.uk>
  *
  *	This program is free software; you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -252,7 +252,7 @@ typedef struct {
 
 typedef struct {
 	DYNARRAY (struct TAG_tnode *, crosses);		/* where these things can be remembered (collects names) */
-	DYNARRAY (int, cross_lexlevels);		/* lexical levels for each of the above */
+	DYNARRAY (int64_t, cross_lexlevels);		/* lexical levels for each of the above */
 	struct TAG_ntdef *resolve_nametype_first;	/* when resolving names, look for these [nodetype] first */
 } guppy_scope_t;
